home-cli/pyproject.toml
[project]
name = "home-cli"
version = "0.1.0"
description = "A CLI tool for controlling your Home Assistant smart home"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.15",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.scripts]
home = "home_cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/home_cli"]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-httpx>=0.35",
    "ruff>=0.9",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
